booking_hotel_info layout fixed, margin decoration moved to uiKit module
