+++ feature/booking/src/main/java/com/zhigaras/booking/ui/adapters/HotelInfoAdapter.kt
@@ -5,12 +5,12 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.zhigaras.adapterdelegate.DelegateAdapter
 import com.zhigaras.adapterdelegate.DelegateViewHolder
+import com.zhigaras.booking.databinding.BookingHotelInfoBinding
 import com.zhigaras.booking.ui.model.HotelInfoUiModel
-import com.zhigaras.uikit.databinding.HotelInfoLayoutBinding
 
 class HotelInfoAdapter : DelegateAdapter<HotelInfoUiModel, HotelInfoAdapter.HotelInfoViewHolder>() {
     
-    inner class HotelInfoViewHolder(private val binding: HotelInfoLayoutBinding) :
+    inner class HotelInfoViewHolder(private val binding: BookingHotelInfoBinding) :
         DelegateViewHolder<HotelInfoUiModel>(binding) {
         
         override fun bind(item: HotelInfoUiModel) {
@@ -21,6 +21,6 @@ class HotelInfoAdapter : DelegateAdapter<HotelInfoUiModel, HotelInfoAdapter.Hote
     override fun viewType() = HotelInfoUiModel::class.hashCode()
     
     override fun createViewHolder(parent: ViewGroup): RecyclerView.ViewHolder = HotelInfoViewHolder(
-        HotelInfoLayoutBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        BookingHotelInfoBinding.inflate(LayoutInflater.from(parent.context), parent, false)
     )
 }
