+++ feature/booking/src/main/res/layout/booking_hotel_info.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@drawable/common_card_bg"
+    android:padding="16dp">
+
+    <include
+        android:id="@+id/hotel_info_layout"
+        layout="@layout/hotel_info_layout" />
+
+</FrameLayout>
