+++ feature/hotel/src/main/res/layout/about_hotel.xml
@@ -26,7 +26,8 @@
         android:id="@+id/hotel_info"
         layout="@layout/hotel_info_layout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
+        android:layout_height="wrap_content"
+        android:layout_marginVertical="16dp" />
 
     <include
         android:id="@+id/price_view"
