+++ uiKit/src/main/res/layout/hotel_info_layout.xml
@@ -4,7 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    android:paddingVertical="16dp"
     tools:parentTag="android.widget.LinearLayout">
 
     <androidx.appcompat.widget.AppCompatTextView
